requirements.txt
aiogram==3.1.1
paho-mqtt==1.6.1